Move the edit board game box modal to the board game box details page.
