+++ src/app/pages/board-game-boxes/board-game-boxes.component.html
@@ -75,91 +75,6 @@
   </div>
 
 
-  <!-- Edit Board Game Box Modal -->
-  <div *ngIf="showEditBoardGameBoxModal" class="modal-overlay" (click)="closeEditBoardGameBoxModal()">
-    <div class="modal-content" (click)="$event.stopPropagation()">
-      <div class="modal-header">
-        <h2>Update Board Game Box</h2>
-        <button type="button" class="close-button" (click)="closeEditBoardGameBoxModal()">×</button>
-      </div>
-      
-      <form (ngSubmit)="onSubmitEditBoardGameBox()" #editBoardGameBoxForm="ngForm" class="modal-form">
-        <div class="form-group">
-          <label for="editTitle">Title *</label>
-          <app-selectable-text-input
-            #titleField
-            id="editTitle"
-            name="editTitle"
-            [(ngModel)]="editBoardGameBox.title"
-            #editTitleNgModel="ngModel"
-            [required]="true"
-            [minlength]="1"
-            placeholder="">
-          </app-selectable-text-input>
-          <div *ngIf="editTitleNgModel.invalid && editTitleNgModel.touched" class="error-text">
-            Title is required
-          </div>
-        </div>
-
-        <div class="form-group">
-          <app-custom-checkbox
-            id="editIsExpansion"
-            name="editIsExpansion"
-            label="Expansion"
-            [(ngModel)]="editBoardGameBox.isExpansion"
-            (ngModelChange)="onExpansionChange($event, true)">
-          </app-custom-checkbox>
-        </div>
-
-        <div class="form-group">
-          <app-custom-checkbox
-            id="editIsStandAlone"
-            name="editIsStandAlone"
-            label="Stand Alone"
-            [(ngModel)]="editBoardGameBox.isStandAlone">
-          </app-custom-checkbox>
-        </div>
-
-        <div class="form-group" *ngIf="editBoardGameBox.isExpansion">
-          <label for="editBaseSetId">Base Set</label>
-          <app-filterable-dropdown
-            id="editBaseSetId"
-            name="editBaseSetId"
-            [(ngModel)]="editBoardGameBox.baseSetId"
-            [options]="baseSetOptions"
-            placeholder="Select a base set (optional)">
-          </app-filterable-dropdown>
-        </div>
-
-        <div class="form-group">
-          <label for="editBoardGameId">Related Board Game</label>
-          <app-filterable-dropdown
-            id="editBoardGameId"
-            name="editBoardGameId"
-            [(ngModel)]="editBoardGameBox.boardGameId"
-            [options]="boardGameOptions"
-            placeholder="">
-          </app-filterable-dropdown>
-        </div>
-
-        <!-- Dynamic Custom Fields -->
-        <app-dynamic-custom-fields 
-          entityKey="boardGameBox" 
-          sectionTitle=""
-          [(customFieldValues)]="editBoardGameBox.customFieldValues">
-        </app-dynamic-custom-fields>
-
-        <div class="modal-actions">
-          <button type="button" class="cancel-button" (click)="closeEditBoardGameBoxModal()">
-            Cancel
-          </button>
-          <button type="submit" class="create-button" [disabled]="!editBoardGameBoxForm.form.valid || isUpdating">
-            {{isUpdating ? 'Updating...' : 'Update'}}
-          </button>
-        </div>
-      </form>
-    </div>
-  </div>
 
   <!-- New Board Game Box Modal -->
   <div *ngIf="showNewBoardGameBoxModal" class="modal-overlay" (click)="closeNewBoardGameBoxModal()">
